chore: reorganize package structure for content stats updater
