+++ application/src/main/java/run/halo/app/content/comment/CommentServiceImpl.java
@@ -16,6 +16,7 @@ import org.springframework.util.Assert;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.util.retry.Retry;
+import run.halo.app.core.counter.CounterService;
 import run.halo.app.core.extension.content.Comment;
 import run.halo.app.core.user.service.RoleService;
 import run.halo.app.core.user.service.UserService;
@@ -29,7 +30,6 @@ import run.halo.app.extension.Ref;
 import run.halo.app.extension.router.selector.FieldSelector;
 import run.halo.app.infra.SystemConfigurableEnvironmentFetcher;
 import run.halo.app.infra.exception.AccessDeniedException;
-import run.halo.app.metrics.CounterService;
 import run.halo.app.plugin.extensionpoint.ExtensionGetter;
 
 /**
